+++ packages/plugin-rows/src/editor/row.tsx
@@ -1,5 +1,5 @@
 import { Plugin, StatefulPluginEditorProps, selectors } from '@edtr-io/core'
-import { OnClickOutside } from '@edtr-io/editor-ui'
+import { OnClickOutside, PrimarySettings } from '@edtr-io/editor-ui'
 import { ThemeProvider, usePluginTheme } from '@edtr-io/ui'
 import * as React from 'react'
 import { createPortal } from 'react-dom'
@@ -7,16 +7,21 @@ import { ReactReduxContextValue } from 'react-redux'
 
 import { rowsPluginThemeFactory, rowsState } from '..'
 import { RowContainer } from '../row-container'
-import {
-  Controls,
-  ExtendedSettingsWrapper,
-  createPrimarySettingsWrapper
-} from './controls'
+import { Controls, ExtendedSettingsWrapper } from './controls'
 import { connectDnD, CollectedProps, TargetProps } from './dnd-hoc'
 import { Menu } from './menu'
 import render from './render'
 import { Separator } from './separator'
 
+const PrimarySettingsWrapper: React.FunctionComponent = props => {
+  React.useEffect(() => {
+    // eslint-disable-next-line no-console
+    console.warn(
+      'PrimarySettingsWrapper of @edtr-io/plugin-rows is deprecated. Use PrimarySettings of @edtr-io/editor-ui instead.'
+    )
+  }, [])
+  return <PrimarySettings {...props} />
+}
 export type RowSourceProps = StatefulPluginEditorProps<typeof rowsState> &
   CollectedProps &
   TargetProps & {
@@ -145,9 +150,7 @@ const RowSource = React.forwardRef<
               extendedSettingsNode.current
             )
           },
-          PrimarySettingsWrapper: createPrimarySettingsWrapper({
-            expanded
-          })
+          PrimarySettingsWrapper
         })}
         <ExtendedSettingsWrapper
           hideExtendedSettings={() => {