+++ packages/plugin-video/src/editor.tsx
@@ -1,5 +1,5 @@
 import { StatefulPluginEditorProps } from '@edtr-io/core'
-import { EditorInput } from '@edtr-io/editor-ui'
+import { EditorInput, PrimarySettings } from '@edtr-io/editor-ui'
 import * as React from 'react'
 
 import { VideoRenderer } from './renderer'
@@ -13,15 +13,17 @@ export const VideoEditor = (
     <React.Fragment>
       <VideoRenderer {...props} disableCursorEvents={editable} />
       {focused ? (
-        <EditorInput
-          label="Video URL:"
-          value={state.value}
-          onChange={e => {
-            state.set(e.target.value)
-          }}
-          textfieldWidth="80%"
-          editorInputWidth="100%"
-        />
+        <PrimarySettings>
+          <EditorInput
+            label="Video URL:"
+            value={state.value}
+            onChange={e => {
+              state.set(e.target.value)
+            }}
+            textfieldWidth="80%"
+            editorInputWidth="100%"
+          />
+        </PrimarySettings>
       ) : null}
     </React.Fragment>
   )